// api-sankhya-center/src/dicionario/tddins/models/tddins.interface.ts
export interface Tddins {
  nuinstancia: number
  nometab: string
  nomeinstancia: string
  descrinstancia: string
  raiz: string
  filtro: string
  ativo: string
  expressao?: string
  nuinstanciapai?: number
  nomescriptchave?: string
  nuinstanciaext?: number
  adicional?: string
  resourceid?: string
  definicaoinst: string
  islib: string
  controle?: string
  descrtela?: string
  categoria?: string
  tipoform?: string
  domain?: string
}